feat(attestations): implement attestationvalidator
Implements the AttestationValidator class to perform a deeper validation
EAS attestations according to our spec. Attestation should at least
contain a chainID and contractAddress for a chain and contract that we
support. The tokenID in the attestaion should point to an hypercert
claimID.

To support these schemas, the SchemaValidator has been split into an Ajv
and Zod validator. Additionally, utils/tokenIds.ts has been added to
validate the value of a tokenID.

Tests have been updated accordingly.